requirements.txt
streamlit
pandas
numpy
rapidfuzz
openai
fpdf2
beautifulsoup4
lxml
pillow
pytesseract
plotly
requests
trafilatura
shap
matplotlib